+++ native/gpu/src/index.d.ts
@@ -3036,8 +3036,6 @@ interface Renderer {
   getCurrentTextureView(): GPUTextureView
   swap()
   waitForPresent(): Promise<boolean>
-  setModalFrameCallback(callback?: () => void): void
-  setModalStateCallback(callback?: (active: boolean) => void): void
   resize()
 }