+++ src/pixi-node/nativeTypes.ts
@@ -18,8 +18,6 @@ export interface NodeWindowRenderer {
   getCurrentTextureView(): GPUTextureView;
   swap(): void;
   waitForPresent?(): Promise<boolean>;
-  setModalFrameCallback?(callback?: () => void): void;
-  setModalStateCallback?(callback?: (active: boolean) => void): void;
   resize(): void;
   destroy(): void;
 }